+++ lib/zego_uikit_signaling_plugin.dart
@@ -1,25 +1,35 @@
 // Dart imports:
 import 'dart:async';
+
+// Flutter imports:
+import 'package:flutter/foundation.dart' show debugPrint;
+import 'package:flutter/services.dart';
+
+// Package imports:
+import 'package:zego_plugin_adapter/zego_plugin_adapter.dart';
+import 'package:zego_zim/zego_zim.dart';
+import 'package:zego_zpns/zego_zpns.dart';
+
+// Project imports:
+import 'package:zego_uikit_signaling_plugin/src/internal/zego_signaling_plugin_event_center.dart';
+
 import 'dart:io'
     if (dart.library.html) 'dart:html'
     if (dart.library.io) 'dart:io';
 
-// Flutter imports:
-import 'package:flutter/foundation.dart' show debugPrint, kIsWeb;
-import 'package:flutter/services.dart';
-// Package imports:
-import 'package:zego_plugin_adapter/zego_plugin_adapter.dart';
-// Project imports:
-import 'package:zego_uikit_signaling_plugin/src/internal/zego_signaling_plugin_event_center.dart';
-import 'package:zego_zim/zego_zim.dart';
-import 'package:zego_zpns/zego_zpns.dart';
+
+
 
 export 'package:zego_zim/zego_zim.dart' hide ZIMEventHandler;
 
 part 'src/zego_signaling_plugin_invitation.dart';
+
 part 'src/zego_signaling_plugin_message.dart';
+
 part 'src/zego_signaling_plugin_notification.dart';
+
 part 'src/zego_signaling_plugin_room.dart';
+
 part 'src/zego_signaling_plugin_user.dart';
 
 class ZegoUIKitSignalingPlugin
@@ -37,7 +47,9 @@ class ZegoUIKitSignalingPlugin
         IZegoUIKitPlugin
     implements ZegoSignalingPluginInterface {
   factory ZegoUIKitSignalingPlugin() => instance;
+
   ZegoUIKitSignalingPlugin._();
+
   static final ZegoUIKitSignalingPlugin instance = ZegoUIKitSignalingPlugin._();
 
   @override
@@ -46,7 +58,7 @@ class ZegoUIKitSignalingPlugin
   @override
   Future<String> getVersion() async {
     final zimVersion = await ZIM.getVersion();
-    if ((!kIsWeb) && (Platform.isAndroid || Platform.isIOS)) {
+    if (Platform.isAndroid || Platform.isIOS) {
       final zpnsVersion = await ZPNs.getVersion();
       return 'signaling:2.0.0;zim:$zimVersion;zpns:$zpnsVersion;';
     } else {